renamed folder to FormFields, reformated dropdown component's placeholder
