+++ src/pages/form_testing.tsx
@@ -1,10 +1,10 @@
 import * as React from "react";
 import { useForm } from "react-hook-form";
 
-import DropDown from "@/components/Form Fields/dropdown-headless";
-import Field_Input from "@/components/Form Fields/field-input";
-import Field_Radio from "@/components/Form Fields/field-radio";
-import Field_Text_Area from "@/components/Form Fields/field-text-area";
+import DropDown from "@/components/FormFields/dropdown-headless";
+import Field_Input from "@/components/FormFields/field-input";
+import Field_Radio from "@/components/FormFields/field-radio";
+import Field_Text_Area from "@/components/FormFields/field-text-area";
 
 /**
  * This is a sample form page using React Hook Form for testing purposes
